Fixed problem with walking animation from endboss
In the if condition where the endboss goes into alert mode is when difference in x coordinate betweend character and endboss is less than 700. When x difference is less than 500 the walking animation fires. The problem was that i only checked is it less than 700 and is it less than 500. For the alert function it was also necessary to check that x difference is below 700 and above 500 so its in a fixed range of x = 700 - 500 = 200 in which only the alert animation fires. Less than 500 only the walking animation fires.